move APIFnVerifySignature to action_verifysignature.go
